requirements.txt
notion-client==2.2.1
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
python-dotenv
requests
